Extract the two Tier-1 LOC-cap hits in userspace-dp/src/afxdp/frame/mod.rs into per-address-family submodules under frame/build/ and frame/rewrite/:
- build_forwarded_frame_into_from_frame (236 LOC body) → split
  into frame/build/mod.rs (orchestrator, ~165 LOC including
  debug-log blocks) + frame/build/ipv4.rs (84 LOC) +
  frame/build/ipv6.rs (65 LOC).
- apply_rewrite_descriptor (223 LOC body) → split into
  frame/rewrite/mod.rs (orchestrator, ~80 LOC) +
  frame/rewrite/ipv4.rs (123 LOC) + frame/rewrite/ipv6.rs
  (103 LOC).

Pure code motion. Per-family bodies byte-identical to the
existing AF_INET / AF_INET6 / 0x0800 / 0x86dd arms.

frame/mod.rs drops from 1783 prod LOC to 1326 LOC (well below
the 1500-LOC Tier-2 hint).

Codegen contract (see docs/pr/1352-frame-build-rewrite-split/plan.md):

- #[inline(always)] on per-family helpers
  (build_forwarded_frame_into_ipv4/6,
  apply_rewrite_descriptor_ipv4/6): each helper has exactly
  ONE call site (its orchestrator's match arm), so
  #[inline(always)] carries zero i-cache duplication risk and
  bypasses the SysV AMD64 ABI register-spill that the
  8-parameter signature would otherwise cause.
- #[inline(never)] + concrete meta: ForwardPacketMeta on
  build_forwarded_frame_into_from_frame: 5 caller wrapper
  sites (frame/mod.rs:201/470, tx/dispatch.rs:505/651/785)
  ⇒ transitive duplication via #[inline(always)] helper
  monomorphizations is a real risk. #[inline(never)] keeps
  ONE binary definition; concrete meta closes the
  generic-monomorphization axis. The wrappers stay generic
  (impl Into<ForwardPacketMeta>) and convert via .into()
  before forwarding to the concrete orchestrator.
- Standard #[inline] on apply_rewrite_descriptor: exactly
  ONE caller (poll_descriptor.rs:746) inside the per-packet
  SIMD descriptor loop ⇒ transitive duplication impossible;
  letting LLVM heuristic-inline restores inter-procedural
  register allocation across the loop.

Plan + adversarial review:

- Plan doc: docs/pr/1352-frame-build-rewrite-split/plan.md
- 6 rounds of Codex + AGY adversarial PLAN review;
  reviewer-ids tracked in docs/pr/1352-frame-build-rewrite-split/reviewer-ids.md.
- Final: Codex r6 = PLAN-NEEDS-MINOR (3 trivial text fixes),
  AGY r6 = PLAN-READY. Plan v5-final.1 closed the 3 Codex
  minors.

Validation:

- cargo build --release clean.
- 1487 cargo tests pass (113 frame-specific tests all pass).
- Codegen gate (nm -C target/release/xpf-userspace-dp):
  * ZERO per-family helper definitions
    (build_forwarded_frame_into_ipv4/6,
    apply_rewrite_descriptor_ipv4/6) ⇒ #[inline(always)]
    fully folded into orchestrators.
  * Exactly ONE definition of
    build_forwarded_frame_into_from_frame ⇒ #[inline(never)]
    + concrete meta worked; no transitive duplication.
  * ZERO definitions of apply_rewrite_descriptor ⇒ LLVM
    inlined the body fully into the single caller at
    poll_descriptor.rs:746 (acceptable per the v5 gate; ZERO
    or ONE both pass).

One pre-existing cross-worktree doc-guard test failure
(snat_contract_documents_current_fail_closed_runtime) is
unrelated to this refactor; the test panics with a path
pointing at a sibling worktree.

Closes #1352
