+++ docs/pr/1352-frame-build-rewrite-split/reviewer-ids.md
@@ -1,0 +1,34 @@
+# #1352 reviewer task IDs
+
+## Plan v1 round 1 (SHA 6ea785c5)
+
+- Codex: `task-mpmyvdz3-7x6rhd`
+- AGY:   `adversarial-review-mpmz4yjh-gerhrw`
+- Claude SMR: in-conversation (this session)
+
+## Plan v2 round 2 (SHA f3b6f482)
+
+- Codex: `task-mpmzar03-0i9jmr`
+- AGY:   `adversarial-review-mpmzb3y8-hxt9tj`
+
+## Plan v3 round 3 (SHA e21e879e)
+
+- Codex: `task-mpmzkrop-r368rl`
+- AGY:   `adversarial-review-mpmzl98h-i50fex`
+
+## Plan v4 round 4 (SHA e64f8f3b)
+
+- Codex: `task-mpmzw5uv-456ix7`
+- AGY:   `adversarial-review-mpmzwsfm-wku4m6`
+
+## Plan v5 round 5 (SHA d810a146)
+
+- Codex: `task-mpn05k1i-ptsc2u`
+- AGY:   `adversarial-review-mpn05vgi-v5hznv`
+
+## Plan v5-final round 6 (SHA 04879451)
+
+- Codex: `task-mpn0de57-aeipws`
+- AGY:   `adversarial-review-mpn0dmc8-fka8d5`
+
+
