+++ userspace-dp/src/afxdp/frame/mod.rs
@@ -68,7 +68,8 @@ pub(in crate::afxdp) use tcp::{
     build_syn_cookie_ack_rst_frame, build_syn_cookie_syn_ack_frame, frame_has_tcp_rst,
 };
 pub(super) use tcp::{extract_tcp_flags_and_window, extract_tcp_window, tcp_flags_str};
-use tcp::clamp_tcp_mss_frame;
+// #1352: clamp_tcp_mss_frame is now imported directly by the per-family
+// helpers in frame/build/{ipv4,ipv6}.rs.
 
 // #1046: TCP segmentation builders extracted into tcp_segmentation.rs
 // to keep frame/mod.rs under the modularity-discipline LOC threshold.
@@ -78,6 +79,18 @@ mod tcp_segmentation;
 pub(in crate::afxdp) use tcp_segmentation::{
     segment_forwarded_tcp_frames, segment_forwarded_tcp_frames_from_frame,
 };
+
+// #1352 — extracted build orchestrator + per-address-family helpers.
+// See docs/pr/1352-frame-build-rewrite-split/plan.md.
+// Wrappers (build_forwarded_frame_from_frame, build_forwarded_frame_into)
+// stay in this file and forward into the orchestrator after
+// converting the meta type via .into().
+mod build;
+pub(in crate::afxdp) use build::build_forwarded_frame_into_from_frame;
+
+// #1352 — extracted apply_rewrite_descriptor orchestrator + per-AF helpers.
+mod rewrite;
+pub(in crate::afxdp) use rewrite::apply_rewrite_descriptor;
 
 
 
@@ -227,242 +240,6 @@ pub(super) fn build_forwarded_frame_from_frame(
 }
 
 
-pub(super) fn build_forwarded_frame_into_from_frame(
-    out: &mut [u8],
-    frame: &[u8],
-    meta: impl Into<ForwardPacketMeta>,
-    decision: &SessionDecision,
-    forwarding: &ForwardingState,
-    apply_nat_on_fabric: bool,
-    expected_ports: Option<(u16, u16)>,
-) -> Option<usize> {
-    let meta = meta.into();
-    let dst_mac = decision.resolution.neighbor_mac?;
-    let enforced_ports = expected_ports;
-    // Use meta L3 offset when it's a valid Ethernet header size (14 or 18),
-    // otherwise re-derive from the frame's ethertype.
-    let l3 = match meta.l3_offset {
-        14 | 18 => meta.l3_offset as usize,
-        _ => frame_l3_offset(frame)?,
-    };
-    if l3 >= frame.len() {
-        return None;
-    }
-    let raw_payload = &frame[l3..];
-    let payload = trim_l3_payload(raw_payload, meta);
-    let (src_mac, vlan_id, apply_nat) =
-        if decision.resolution.disposition == ForwardingDisposition::FabricRedirect {
-            (
-                decision.resolution.src_mac?,
-                decision.resolution.tx_vlan_id,
-                apply_nat_on_fabric,
-            )
-        } else {
-            (
-                decision.resolution.src_mac?,
-                decision.resolution.tx_vlan_id,
-                true,
-            )
-        };
-    let eth_len = if vlan_id > 0 { 18 } else { 14 };
-    let ether_type = match meta.addr_family as i32 {
-        libc::AF_INET => 0x0800,
-        libc::AF_INET6 => 0x86dd,
-        _ => return None,
-    };
-    let frame_len = eth_len + payload.len();
-    if frame_len > out.len() {
-        return None;
-    }
-    write_eth_header_slice(
-        out.get_mut(..eth_len)?,
-        dst_mac,
-        src_mac,
-        vlan_id,
-        ether_type,
-    )?;
-    let payload_out = out.get_mut(eth_len..frame_len)?;
-    // SAFETY: source (payload) and destination (payload_out) are distinct
-    // buffers — payload is from the ingress UMEM, payload_out is in the
-    // egress UMEM. Lengths are equal because both span eth_len..frame_len.
-    debug_assert_eq!(payload_out.len(), payload.len());
-    unsafe {
-        core::ptr::copy_nonoverlapping(payload.as_ptr(), payload_out.as_mut_ptr(), payload.len());
-    }
-    let out = &mut out[..frame_len];
-    let force_tunnel_l4_recompute = decision.resolution.tunnel_endpoint_id != 0;
-    let tunnel_tcp_mss = native_gre_tcp_mss(forwarding, decision, meta.addr_family);
-    let ip_start = eth_len;
-    match meta.addr_family as i32 {
-        libc::AF_INET => {
-            if out.len() < ip_start + 20 {
-                return None;
-            }
-            let ihl = ((out[ip_start] & 0x0f) as usize) * 4;
-            if ihl < 20 || out.len() < ip_start + ihl {
-                return None;
-            }
-            if (meta.meta_flags & 0x80) == 0 && out[ip_start + 8] <= 1 {
-                return None;
-            }
-            let old_src = Ipv4Addr::new(
-                out[ip_start + 12],
-                out[ip_start + 13],
-                out[ip_start + 14],
-                out[ip_start + 15],
-            );
-            let old_dst = Ipv4Addr::new(
-                out[ip_start + 16],
-                out[ip_start + 17],
-                out[ip_start + 18],
-                out[ip_start + 19],
-            );
-            let old_ttl = out[ip_start + 8];
-            // IHL already computed above — use directly instead of re-parsing.
-            let rel_l4 = ihl;
-            let repaired_ports =
-                restore_l4_tuple_from_meta(&mut out[ip_start..], meta, rel_l4).unwrap_or(false);
-            if apply_nat {
-                apply_nat_ipv4(&mut out[ip_start..], meta.protocol, decision.nat)?;
-            }
-            let skip_ttl = (meta.meta_flags & 0x80) != 0;
-            if !skip_ttl {
-                out[ip_start + 8] -= 1;
-            }
-            let enforced = enforce_expected_ports_at(
-                out,
-                ip_start,
-                ip_start + rel_l4,
-                meta.addr_family,
-                meta.protocol,
-                enforced_ports,
-            )
-            .unwrap_or(false);
-            adjust_ipv4_header_checksum(
-                &mut out[ip_start..ip_start + ihl],
-                old_src,
-                old_dst,
-                old_ttl,
-            )?;
-            if tunnel_tcp_mss > 0 {
-                let _ = clamp_tcp_mss_frame(out, ip_start, tunnel_tcp_mss);
-            }
-            if force_tunnel_l4_recompute || (repaired_ports && !enforced) {
-                recompute_l4_checksum_ipv4(&mut out[ip_start..], ihl, meta.protocol, true)?;
-            }
-        }
-        libc::AF_INET6 => {
-            if out.len() < ip_start + 40 {
-                return None;
-            }
-            if (meta.meta_flags & 0x80) == 0 && out[ip_start + 7] <= 1 {
-                return None;
-            }
-            // Use meta-derived L4 offset when valid (>= 40 for IPv6 base header,
-            // avoids walking extension headers). Fall back to parsing otherwise.
-            let meta_rel = meta.l4_offset.wrapping_sub(meta.l3_offset) as usize;
-            let rel_l4 = if meta_rel >= 40 && meta.l4_offset > meta.l3_offset {
-                meta_rel
-            } else {
-                packet_rel_l4_offset(&out[ip_start..], meta.addr_family)?
-            };
-            let repaired_ports =
-                restore_l4_tuple_from_meta(&mut out[ip_start..], meta, rel_l4).unwrap_or(false);
-            if apply_nat {
-                apply_nat_ipv6(&mut out[ip_start..], meta.protocol, decision.nat)?;
-            }
-            if (meta.meta_flags & 0x80) == 0 {
-                out[ip_start + 7] -= 1;
-            }
-            let enforced = enforce_expected_ports_at(
-                out,
-                ip_start,
-                ip_start + rel_l4,
-                meta.addr_family,
-                meta.protocol,
-                enforced_ports,
-            )
-            .unwrap_or(false);
-            if tunnel_tcp_mss > 0 {
-                let _ = clamp_tcp_mss_frame(out, ip_start, tunnel_tcp_mss);
-            }
-            if force_tunnel_l4_recompute || (repaired_ports && !enforced) {
-                recompute_l4_checksum_ipv6(&mut out[ip_start..], meta.protocol)?;
-            }
-        }
-        _ => return None,
-    }
-    // Debug: dump first N built frames' Ethernet + IP headers to see post-NAT on wire
-    if cfg!(feature = "debug-log") {
-        thread_local! {
-            static BUILD_FWD_DBG_COUNT: std::cell::Cell<u32> = const { std::cell::Cell::new(0) };
-        }
-        BUILD_FWD_DBG_COUNT.with(|c| {
-            let n = c.get();
-            if n < 30 {
-                c.set(n + 1);
-                let pkt_detail = decode_frame_summary(out);
-                eprintln!(
-                    "DBG BUILT_ETH[{}]: vlan={} frame_len={} proto={} {}",
-                    n, vlan_id, frame_len, meta.protocol, pkt_detail,
-                );
-                // For the first 3 frames, also dump the full IP+TCP header hex
-                if n < 3 {
-                    let dump_len = frame_len.min(out.len()).min(eth_len + 60);
-                    let hex: String = out[..dump_len]
-                        .iter()
-                        .map(|b| format!("{:02x}", b))
-                        .collect::<Vec<_>>()
-                        .join(" ");
-                    eprintln!("DBG BUILT_HEX[{n}]: {hex}");
-                }
-            }
-        });
-    }
-    // Checksum verification: recompute from scratch and compare to incremental update.
-    if cfg!(feature = "debug-log") {
-        verify_built_frame_checksums(&out[..frame_len]);
-    }
-
-    // RST corruption check: detect if frame building introduced a TCP RST
-    // that wasn't in the source frame.
-    if cfg!(feature = "debug-log") {
-        let out_has_rst = frame_has_tcp_rst(&out[..frame_len]);
-        let in_has_rst = frame_has_tcp_rst(frame);
-        if out_has_rst && !in_has_rst {
-            thread_local! {
-                static BUILD_RST_CORRUPT_COUNT: std::cell::Cell<u32> = const { std::cell::Cell::new(0) };
-            }
-            BUILD_RST_CORRUPT_COUNT.with(|c| {
-                let n = c.get();
-                if n < 20 {
-                    c.set(n + 1);
-                    let in_summary = decode_frame_summary(frame);
-                    let out_summary = decode_frame_summary(&out[..frame_len]);
-                    eprintln!(
-                        "RST_CORRUPT BUILD[{}]: frame build INTRODUCED RST! in=[{}] out=[{}]",
-                        n, in_summary, out_summary,
-                    );
-                    let in_hex_len = frame.len().min(80);
-                    let in_hex: String = frame[..in_hex_len]
-                        .iter()
-                        .map(|b| format!("{:02x}", b))
-                        .collect::<Vec<_>>()
-                        .join(" ");
-                    let out_hex_len = frame_len.min(out.len()).min(80);
-                    let out_hex: String = out[..out_hex_len]
-                        .iter()
-                        .map(|b| format!("{:02x}", b))
-                        .collect::<Vec<_>>()
-                        .join(" ");
-                    eprintln!("RST_CORRUPT IN_HEX[{n}]: {in_hex}");
-                    eprintln!("RST_CORRUPT OUT_HEX[{n}]: {out_hex}");
-                }
-            });
-        }
-    }
-    Some(frame_len)
-}
 
 #[cfg_attr(not(test), allow(dead_code))]
 pub(super) fn build_forwarded_frame(
@@ -489,6 +266,8 @@ pub(super) fn build_forwarded_frame_into(
     expected_ports: Option<(u16, u16)>,
 ) -> Option<usize> {
     let frame = area.slice(desc.addr as usize, desc.len as usize)?;
+    // v5: orchestrator takes concrete ForwardPacketMeta (#1352 plan).
+    let meta = meta.into();
     build_forwarded_frame_into_from_frame(
         out,
         frame,
@@ -902,242 +681,6 @@ fn trim_l3_payload<'a>(raw_payload: &'a [u8], meta: impl Into<ForwardPacketMeta>
     }
 }
 
-/// Straight-line frame rewrite using a precomputed `RewriteDescriptor`.
-///
-/// Eliminates per-packet branches for address family, VLAN presence, NAT type,
-/// and checksum recomputation — all decisions are baked into the descriptor at
-/// session / flow-cache insertion time.
-///
-/// Returns the new frame length on success, or `None` if the frame is corrupt,
-/// too short, or has a port mismatch (caller falls back to generic rewrite).
-///
-/// **Scope**: IPv4/IPv6 TCP and UDP only (flow cache gates on ACK-only TCP + UDP).
-/// Does NOT handle: ICMP identifier repair, NAT64 (header-size change), NPTv6
-/// (checksum-neutral — no L4 csum adjust needed, but address rewrite differs).
-#[inline]
-pub(super) fn apply_rewrite_descriptor(
-    area: &MmapArea,
-    desc: XdpDesc,
-    meta: UserspaceDpMeta,
-    rd: &super::RewriteDescriptor,
-    expected_ports: Option<(u16, u16)>,
-) -> Option<InPlaceRewriteResult> {
-    // NAT64 and NPTv6 use the generic path — they need special handling.
-    if rd.nat64 || rd.nptv6 {
-        return None;
-    }
-
-    let prep = rewrite_prepare_eth_from_parts(
-        area,
-        desc,
-        meta.into(),
-        RewriteEthParams {
-            dst_mac: rd.dst_mac,
-            src_mac: rd.src_mac,
-            vlan_id: rd.tx_vlan_id,
-            ether_type: rd.ether_type,
-            apply_nat: !rd.fabric_redirect || rd.apply_nat_on_fabric,
-        },
-    )?;
-    let packet = unsafe { area.slice_mut_unchecked(prep.tx_offset as usize, prep.frame_len)? };
-    let frame_len = prep.frame_len;
-    let ip = prep.ip_start;
-    let skip_ttl = prep.skip_ttl;
-    let apply_nat = prep.apply_nat;
-
-    match rd.ether_type {
-        0x0800 => {
-            // ── IPv4 straight-line rewrite ──
-            if packet.len() < ip + 20 {
-                return None;
-            }
-            let ihl = ((packet[ip] & 0x0f) as usize) * 4;
-            if ihl < 20 || packet.len() < ip + ihl {
-                return None;
-            }
-            if !skip_ttl && packet[ip + 8] <= 1 {
-                return None; // TTL expired
-            }
-            let l4 = ip + ihl;
-
-            // Port validation (DMA race guard).
-            // If ports don't match, fall back to generic path for repair.
-            if let Some((exp_src, exp_dst)) = expected_ports {
-                if matches!(meta.protocol, PROTO_TCP | PROTO_UDP) && packet.len() >= l4 + 4 {
-                    let cur_src = u16::from_be_bytes([packet[l4], packet[l4 + 1]]);
-                    let cur_dst = u16::from_be_bytes([packet[l4 + 2], packet[l4 + 3]]);
-                    if cur_src != exp_src || cur_dst != exp_dst {
-                        return None;
-                    }
-                }
-            }
-
-            // NAT: direct byte writes for IP addresses (#963 PR-B
-            // helpers). Caller-side `if let Some(IpAddr::V4(_))`
-            // matching keeps conditional logic visible at the call
-            // site and lets the `#[inline(always)]` helpers fold
-            // into a single MOV/MOVB instruction.
-            if apply_nat {
-                if let Some(IpAddr::V4(new_src)) = rd.rewrite_src_ip {
-                    write_ipv4_src(packet, ip, new_src);
-                }
-                if let Some(IpAddr::V4(new_dst)) = rd.rewrite_dst_ip {
-                    write_ipv4_dst(packet, ip, new_dst);
-                }
-            }
-
-            // NAT: direct byte writes for L4 ports.
-            if apply_nat {
-                if let Some(new_sport) = rd.rewrite_src_port {
-                    write_l4_src_port(packet, l4, new_sport);
-                }
-                if let Some(new_dport) = rd.rewrite_dst_port {
-                    write_l4_dst_port(packet, l4, new_dport);
-                }
-            }
-
-            // TTL decrement (skip for fabric-ingress — peer already decremented).
-            if !skip_ttl {
-                packet[ip + 8] -= 1;
-            }
-
-            // IP header checksum: precomputed NAT delta + TTL-1 delta.
-            let old_csum = u16::from_be_bytes([packet[ip + 10], packet[ip + 11]]);
-            let mut sum = (!old_csum as u32) & 0xffff;
-            if apply_nat {
-                sum += rd.ip_csum_delta as u32;
-            }
-            if !skip_ttl {
-                // TTL-1 delta is always 0xFEFF in one's complement arithmetic
-                sum += 0xFEFF;
-            }
-            while (sum >> 16) != 0 {
-                sum = (sum & 0xffff) + (sum >> 16);
-            }
-            let new_csum = !(sum as u16);
-            packet[ip + 10..ip + 12].copy_from_slice(&new_csum.to_be_bytes());
-
-            // L4 checksum: precomputed delta covers IP + port changes.
-            if apply_nat && rd.l4_csum_delta != 0 {
-                let l4_csum_off = match meta.protocol {
-                    PROTO_TCP => l4 + 16,
-                    PROTO_UDP => l4 + 6,
-                    _ => 0,
-                };
-                if l4_csum_off > 0 && packet.len() >= l4_csum_off + 2 {
-                    let old_l4_csum =
-                        u16::from_be_bytes([packet[l4_csum_off], packet[l4_csum_off + 1]]);
-                    // Skip UDP checksum update if zero (no checksum, RFC 768).
-                    if meta.protocol != PROTO_UDP || old_l4_csum != 0 {
-                        let mut l4sum = (!old_l4_csum as u32) & 0xffff;
-                        l4sum += rd.l4_csum_delta as u32;
-                        while (l4sum >> 16) != 0 {
-                            l4sum = (l4sum & 0xffff) + (l4sum >> 16);
-                        }
-                        let new_l4 = !(l4sum as u16);
-                        // UDP: 0x0000 means "no checksum" — use 0xFFFF (RFC 768).
-                        let final_csum = if meta.protocol == PROTO_UDP && new_l4 == 0 {
-                            0xFFFFu16
-                        } else {
-                            new_l4
-                        };
-                        packet[l4_csum_off..l4_csum_off + 2]
-                            .copy_from_slice(&final_csum.to_be_bytes());
-                    }
-                }
-            }
-        }
-        0x86dd => {
-            // ── IPv6 straight-line rewrite ──
-            // No IP header checksum; only L4 pseudo-header changes matter.
-            if packet.len() < ip + 40 {
-                return None;
-            }
-            if !skip_ttl && packet[ip + 7] <= 1 {
-                return None; // Hop limit expired
-            }
-
-            // L4 offset from metadata or by parsing extension headers.
-            let meta_rel = meta.l4_offset.wrapping_sub(meta.l3_offset) as usize;
-            let rel_l4 = if meta_rel >= 40 && meta.l4_offset > meta.l3_offset {
-                meta_rel
-            } else {
-                packet_rel_l4_offset(&packet[ip..], meta.addr_family)?
-            };
-            let l4 = ip + rel_l4;
-
-            // Port validation (DMA race guard).
-            if let Some((exp_src, exp_dst)) = expected_ports {
-                if matches!(meta.protocol, PROTO_TCP | PROTO_UDP) && packet.len() >= l4 + 4 {
-                    let cur_src = u16::from_be_bytes([packet[l4], packet[l4 + 1]]);
-                    let cur_dst = u16::from_be_bytes([packet[l4 + 2], packet[l4 + 3]]);
-                    if cur_src != exp_src || cur_dst != exp_dst {
-                        return None;
-                    }
-                }
-            }
-
-            // NAT: direct byte writes for IPv6 addresses (#963 PR-B).
-            if apply_nat {
-                if let Some(IpAddr::V6(new_src)) = rd.rewrite_src_ip {
-                    write_ipv6_src(packet, ip, new_src);
-                }
-                if let Some(IpAddr::V6(new_dst)) = rd.rewrite_dst_ip {
-                    write_ipv6_dst(packet, ip, new_dst);
-                }
-            }
-
-            // NAT: direct byte writes for L4 ports (#963 PR-B).
-            if apply_nat {
-                if let Some(new_sport) = rd.rewrite_src_port {
-                    write_l4_src_port(packet, l4, new_sport);
-                }
-                if let Some(new_dport) = rd.rewrite_dst_port {
-                    write_l4_dst_port(packet, l4, new_dport);
-                }
-            }
-
-            // Hop limit decrement (skip for fabric-ingress).
-            if !skip_ttl {
-                packet[ip + 7] -= 1;
-            }
-
-            // L4 checksum: precomputed delta covers IPv6 address + port changes.
-            if apply_nat && rd.l4_csum_delta != 0 {
-                let l4_csum_off = match meta.protocol {
-                    PROTO_TCP => l4 + 16,
-                    PROTO_UDP => l4 + 6,
-                    PROTO_ICMPV6 => l4 + 2,
-                    _ => 0,
-                };
-                if l4_csum_off > 0 && packet.len() >= l4_csum_off + 2 {
-                    let old_l4_csum =
-                        u16::from_be_bytes([packet[l4_csum_off], packet[l4_csum_off + 1]]);
-                    let mut l4sum = (!old_l4_csum as u32) & 0xffff;
-                    l4sum += rd.l4_csum_delta as u32;
-                    while (l4sum >> 16) != 0 {
-                        l4sum = (l4sum & 0xffff) + (l4sum >> 16);
-                    }
-                    let new_l4 = !(l4sum as u16);
-                    // IPv6 UDP must have non-zero checksum; use 0xFFFF for all.
-                    let final_csum = if new_l4 == 0 { 0xFFFFu16 } else { new_l4 };
-                    packet[l4_csum_off..l4_csum_off + 2].copy_from_slice(&final_csum.to_be_bytes());
-                }
-            }
-        }
-        _ => return None,
-    }
-
-    // Checksum verification for descriptor path (debug only).
-    if cfg!(feature = "debug-log") {
-        verify_built_frame_checksums(&packet[..frame_len]);
-    }
-    Some(InPlaceRewriteResult {
-        offset: prep.tx_offset,
-        len: frame_len as u32,
-        l2_rewrite: prep.l2_rewrite,
-    })
-}
 
 pub(super) fn apply_nat_ipv4(packet: &mut [u8], protocol: u8, nat: NatDecision) -> Option<()> {
     if nat == NatDecision::default() {
